# mygram-api-reza/docs/swagger.yaml
definitions:
  structs.CommentRequest:
    properties:
      message:
        example: This is a comment
        type: string
      photo_id:
        example: 1
        type: integer
    type: object
  structs.CommentResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      user_id:
        type: integer
    type: object
  structs.DeleteCommentResponse:
    properties:
      message:
        type: string
    type: object
  structs.DeletePhotoResponse:
    properties:
      message:
        type: string
    type: object
...
  structs.EmbeddedPhotoResponse:
    properties:
      caption:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
...
  structs.EmbeddedUserResponse:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
  structs.GetCommentResponse:
    properties:
      Photo:
        $ref: '#/definitions/structs.EmbeddedPhotoResponse'
      User:
        $ref: '#/definitions/structs.EmbeddedUserResponse'
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  structs.GetPhotoResponse:
    properties:
      User:
        $ref: '#/definitions/structs.EmbeddedUserResponse'
      caption:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  structs.PhotoRequest:
    properties:
      caption:
        example: This is my first ever photo uploaded
        type: string
      photo_url:
        example: https://mygram.com/photos/1
        type: string
      title:
        example: My first ever photo uploaded
        type: string
    type: object
...
  structs.UpdateCommentRequest:
    properties:
      message:
        example: This is a comment
        type: string
    type: object
  structs.UpdateCommentResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  structs.UpdatePhotoResponse:
    properties:
      caption:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /comment:
    get:
      description: Get all comments
      operationId: get-all-comments
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.GetCommentResponse'
      summary: Get all comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Post new comment on photo
      operationId: post-comment
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post comment request body json
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/structs.CommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/structs.CommentResponse'
      summary: Post new comment on photo
      tags:
      - comments
  /comment/{commentID}:
    delete:
      description: Delete comment
      operationId: delete-comment
      parameters:
      - description: comments's id
        in: path
        name: commentID
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.DeleteCommentResponse'
      summary: Delete existing comment
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Update comment
      operationId: update-comment
      parameters:
      - description: comments's id
        in: path
        name: commentID
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Edit photo request body json
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/structs.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.UpdateCommentResponse'
      summary: Update existing comment
      tags:
      - comments
  /photo:
    get:
      description: Get all photos
      operationId: get-all-photos
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.GetPhotoResponse'
      summary: Get all photos
      tags:
      - photos
    post:
      consumes:
      - application/json
      description: Post a new photo
      operationId: post-photo
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add photo request body json
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/structs.PhotoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/structs.UpdatePhotoResponse'
      summary: Post a new photo
      tags:
      - photos
  /photo/{photoID}:
    delete:
      description: Delete photo
      operationId: delete-photo
      parameters:
      - description: photo's id
        in: path
        name: photoID
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.DeletePhotoResponse'
      summary: Delete existing photo
      tags:
      - photos
    put:
      consumes:
      - application/json
      description: Update photo data
      operationId: update-photo
      parameters:
      - description: photo's id
        in: path
        name: photoID
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Edit photo request body json
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/structs.PhotoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.UpdatePhotoResponse'
      summary: Update existing photo data
      tags:
      - photos
...
swagger: "2.0"